##merged calc to util ##merged alliance to api
